requirements.txt



fastapi
uvicorn
pandas
joblib
scikit-learn
gspread
google-auth
google-auth-oauthlib
python-multipart
numpy
python-dotenv
